requirement.txt
flask
flask-cors
google-generativeai
google-api-python-client
python-dotenv
gunicorn